simulators._ImageGenerator: replace deprecated use of numpy.float.
numpy.float is deprecated since Numpy 1.20 and is an alias for the
builtin `float`.

+++ microscope/simulators/__init__.py
@@ -63,7 +63,7 @@ class _ImageGenerator:
             self.white,
         )
         self._method_index = 0
-        self._datatypes = (np.uint8, np.uint16, np.float)
+        self._datatypes = (np.uint8, np.uint16, float)
         self._datatype_index = 0
         self._theta = _theta_generator()
         self.numbering = True
